REFERENCIA/OnePunch/requirements.txt

# Communication Services
# Core Framework
# Database
# Document Processing
# Environment
# Google APIs
# LLM Providers
# PayPal
# Server
# Testing
# Utilities
# Vector Database & Embeddings
# Voice
livekit-api>=0.1.0
livekit-agents>=0.7.0
livekit-plugins-openai>=0.1.0
livekit-plugins-silero>=0.1.0
livekit-plugins-noise-cancellation>=0.2.0

# CRM Integration
hubspot-api-client>=8.0.0

Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
Flask==3.0.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
anthropic==0.8.1
bcrypt==4.1.2
cryptography==41.0.7
elevenlabs==0.2.27
eventlet==0.33.3
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
gspread==5.12.0
gunicorn==21.2.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain==0.1.0
openai==1.6.1
openpyxl==3.1.2
paypalrestsdk==1.13.1
pdfminer.six
pdfplumber==0.10.3
pinecone-client==5.0.1
pyjwt==2.8.0
pytest-flask==1.3.0
pytest==7.4.3
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sendgrid==6.11.0
twilio==8.10.0
validators==0.22.0
reportlab==4.0.8
paypal-agent-toolkit>=1.1.0
langgraph>=1.0.0
langchain-openai>=0.1.0
